release 3.0.59-rc source code for php

+++ Services/Vod/V1/Model/CheckMd5DuplicationRequest.php
@@ -46,7 +46,7 @@ class CheckMd5DuplicationRequest implements ModelInterface, ArrayAccess
     protected static $openAPIFormats = [
         'authorization' => null,
         'xSdkDate' => null,
-        'size' => 'int32',
+        'size' => 'int64',
         'md5' => null
     ];
